Implemented UserDetails object and UserDetailsService, added AuthenticationManager bean in SecurityConfig
