+++ src/main/java/com/ibm/test/first_project/data/repositories/UserRepository.java
@@ -2,9 +2,11 @@ package com.ibm.test.first_project.data.repositories;
 
 import com.ibm.test.first_project.data.models.User;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.Optional;
 
+@Transactional
 public interface UserRepository extends JpaRepository<User, Long> {
 
     Optional<User> findByUsername(String username);
