+++ src/main/java/com/ibm/test/first_project/services/impl/AppUserService.java
@@ -1,30 +1,24 @@
 package com.ibm.test.first_project.services.impl;
 
-import com.ibm.test.first_project.data.models.User;
 import com.ibm.test.first_project.data.repositories.UserRepository;
-import com.ibm.test.first_project.utils.CustomUserDetails;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 
-import java.util.Optional;
-
 @Service
 @RequiredArgsConstructor
-public class UserDetailsServiceImpl implements UserDetailsService {
+public class AppUserService implements UserDetailsService {
 
     private final UserRepository userRepository;
 
+    private final String USER_NOT_FOUND = "User with username '%s' not found";
+
     @Override
     public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
-        Optional<User> user = userRepository.findByUsername(username);
+        return userRepository.findByUsername(username)
+                .orElseThrow(() -> new UsernameNotFoundException(String.format(USER_NOT_FOUND, username)));
 
-        if (user.isEmpty()) {
-            throw new UsernameNotFoundException(String.format("User with username '%s' does not exist", user));
-        }
-
-        return new CustomUserDetails(user.get());
     }
 }
